ATM client changes: Reorganized code to allow manual UI refresh, cleaned up code and added proper logging
Changed DisplayMessage function to allow "unskippable" messages that must remain until condition is met.
Example uses are a "loading" message when connecting to the server (not  implemented) or a "invalid card" message (implemented)

There are points where the program will not continue until response from the server, but UI may still need to be updated, added a manual UpdateUI function to address this.

cleaned up code removing things already declared in apis

code is reorganized to allow for manual ui updates