Added reconnection logic and user feedback for all clients
Clients now try to reconnect if they get disconnected or can't reach server
Clients call a onDisconnect function to give proper feedback for the user if a disconnect happens.
All bank requests not check if the client is connected and if not they call the callback, if there is one, passing back nil.