Add heat beat system to server that sends pings to clients
atm API has a timer that resets when it get a "beat" command from the server
if this timer hits 0 it "disconnects" from the sever and passed a "onDisconnect" event up to the parent, allowing each client type to handle this "reconnect" differently if need be.

I only have the User Registration terminal reconnecting on this event, the other two client types still just give user feedback only.

Also, sending pings is not in yet, still trying to figure out how to properly track all connected clients on the server, hence the prints in bankAPI's new "broadcast" function.